% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{balancesheets}
\alias{balancesheets}
\title{Balancesheets of all companies}
\description{
A dataset containing all balancesheets, if available,
 for all companies in the companies.RData file. Among other things,
 stores:
 \itemize{
   \item Cash & Equivalents.
   \item Short Term Investments.
   \item Total Inventory.
   \item Prepaid Expenses.
   \item Total Current Assets.
 }
}
\keyword{data}

