% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{cashflows}
\alias{cashflows}
\title{Cashflow statements of all companies}
\description{
A dataset containing all cashflow statements, if available,
 for all companies in the companies.RData file. Among other things,
 stores:
 \itemize{
   \item Net Income/Starting Line.
   \item Depreciation/Depletion.
   \item Amortization.
   \item Deferred Taxes.
   \item Changes in Working Capital.
 }
}
\keyword{data}

