% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clean_downloads.R
\name{clean_downloads}
\alias{clean_downloads}
\title{Removes all downloaded temporary files for a given data frame of companies.}
\usage{
clean_downloads(x = qmjdata::companies)
}
\arguments{
\item{x}{A data frame of companies. Must have a ticker column.}
}
\description{
This function works to remove the files that get_info and get_prices
temporarily store in order to conserve progress.
Because this data may become irrelvant over time,
clean_downloads removes these files so that get_info and
get_prices may download fresh sets of data for a given
data frame of companies.
}
\details{
The clean_downloads() function will also automatically remove any
temporarily stored data for the S&P 500, with the stock ticker ^GSPC.

Returns a logical vector Where the ith and (i+1)th correspond to
whether or not a temporary financial and/or price file, respectively,
was found and removed for the ith company provided.
The last two indices refer to the S&P 500 temporary data.
}
\examples{
sub_comps <- qmjdata::companies[1:2,]
clean_downloads(sub_comps)
}
\seealso{
\code{\link{get_prices}}

\code{\link{get_info}}
}

