% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/collectmarketdata.R
\name{collectmarketdata}
\alias{collectmarketdata}
\title{collectmarketdata}
\usage{
collectmarketdata()
}
\description{
Reads data from companies.csv and calculates market growth, payouts, safety, and profitability
for later processing.
}

