% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/collectmarketgrowth.R
\name{collectmarketgrowth}
\alias{collectmarketgrowth}
\title{collectmarketgrowth}
\usage{
collectmarketgrowth(x, BS, CF, IS)
}
\arguments{
\item{x}{A dataframe of company names and tickers.}

\item{BS}{A dataframe containing balance sheet information for every company.}

\item{CF}{A dataframe containing cash flow information for every company.}

\item{IS}{A dataframe containing income statement information for every company.}
}
\description{
Given a list of companies (names and tickers), a balance sheet, a cash flow statement,
and an income statement, calculates GPOA, ROE, ROA, CFOA, GMAR, ACC
and determines the z-score of overall growth based on the paper
Quality Minus Junk (Asness et al.) in Appendix page A2.
}

