% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/collectmarketpayout.R
\name{collectmarketpayout}
\alias{collectmarketpayout}
\title{collectmarketpayout}
\usage{
collectmarketpayout(x, BS, IS)
}
\arguments{
\item{x}{A dataframe of company names and tickers.}

\item{BS}{A dataframe containing balance sheet information for every company.}

\item{IS}{A dataframe containing income statement information for every company.}
}
\description{
Given a list of companies (names and tickers), a balance sheet, a cash flow statement,
and an income statement, calculates EISS, DISS, NPOP
and determines the z-score of overall payout based on the paper
Quality Minus Junk (Asness et al.) in Appendix page A3-4.
}

