% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{companies}
\alias{companies}
\title{A list of all companies of interest}
\description{
Stores the names and tickers for all companies in the Russell 3000 Index.
}
\keyword{data}

