% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_companies.R
\name{create_companies}
\alias{create_companies}
\title{Creates a list of Company objects.}
\usage{
create_companies(x, financials, prices)
}
\arguments{
\item{x}{A dataframe of company names and tickers.}

\item{financials}{a formatted data frame containing financial information for the given companies.}

\item{prices}{A dataframe containing the daily market closing prices and returns.}
}
\description{
Creates a list of Company objects, which allow for much greater analysis of
an individual company.
}

