% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_qmjAnalyses.R
\name{create_qmjAnalyses}
\alias{create_qmjAnalyses}
\title{Creates a list of qmjAnalyses.}
\usage{
create_qmjAnalyses(x, financials, prices)
}
\arguments{
\item{financials}{A data frame containing financial information for the given companies.}

\item{prices}{A data frame containing the daily market closing prices and returns.}

\item{companies}{A data frame of company names and tickers.}
}
\description{
Creates a list of qmjAnalyses, which allow for much greater analysis of
individual companies.
}
\examples{
data(companies)
data(financials)
data(prices)
create_qmjAnalyses(companies,financials,prices)
}

