% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_qmjs.R
\name{create_qmjs}
\alias{create_qmjs}
\title{Creates a list of qmjs.}
\usage{
create_qmjs(x, financials, prices)
}
\arguments{
\item{x}{A data frame of company names and tickers.}

\item{financials}{A data frame containing financial information for the given companies.}

\item{prices}{A data frame containing the daily market closing prices and returns.}
}
\description{
Creates a list of qmjs, which allow for much greater analysis of
individual companies.
}
\examples{
data(companies)
data(financials)
data(prices)

sub_comps <- companies[1:5,]
create_qmjs(sub_comps,financials,prices)
}

