% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{dailydata}
\alias{dailydata}
\title{A dataframe of price returns for companies}
\description{
Stores calculated price returns (pret) for companies in companies.RData
}
\keyword{data}

