% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{financials}
\alias{financials}
\title{Financial statements of all companies for the past four years}
\description{
A data frame containing all financial statements (balancesheets,
cashflows, and income statements) for the past four years if available.
Data
}
\details{
Data is gathered from Google Finance. Data set contains:
 \itemize{
   \item CE = Cash & Equivalents.
   \item STI = Short Term Investments.
   \item CSTI = Cash and Short Term Investments
   \item AR = Accounts Receivable - Trade, Net
   \item RE = Receivables - Other

}
}
\keyword{data}

