% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_companies.R
\name{get_companies}
\alias{get_companies}
\title{Returns a specific Company object (Company Portfolio)}
\usage{
get_companies(tickers, portfolios)
}
\arguments{
\item{portfolios}{A list of company portfolios, i.e., Company objects.}

\item{...}{A single or vector}
}
\description{
Given a data frame of companies, a ticker, a data frame with financial statements, and a data frame of prices, creates
a particular company portfolio and returns the result.
}

