% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_company.R
\name{get_company}
\alias{get_company}
\title{Returns a specific Company object (Company Portfolio)}
\usage{
get_company(companies, ticker, financials, prices)
}
\arguments{
\item{companies}{A data frame of companies.}

\item{ticker}{A company ticker as a Character. Must already be present in the companies data frame.}

\item{financials}{a formatted data frame containing financial information for the given companies.}

\item{prices}{A dataframe containing the daily market closing prices and returns.}
}
\description{
Given a data frame of companies, a ticker, a data frame with financial statements, and a data frame of prices, creates
a particular company portfolio and returns the result.
}

