% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_dailydata.R
\name{get_dailydata}
\alias{get_dailydata}
\title{Gets the daily prices and returns of companies for the past five years.}
\usage{
get_dailydata(x)
}
\arguments{
\item{x}{A dataframe of company names and tickers.}
}
\description{
Given a list of companies (names and tickers), writes .csv files for every company
storing price returns.
}
\examples{
x <- data(companies)
\dontrun{get_dailydata(x)}
}

