% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_extrafin.R
\name{get_extrafin}
\alias{get_extrafin}
\title{Grabs beta scores for companies as well as their earnings
before interest, taxes, depreciation, and amortization}
\usage{
get_extrafin()
}
\description{
Retrieves the beta of companies based on
yahoo finance's benchmark.
}
\examples{
\donttest{
get_extrafin()
}
}

