% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_portfolios.R
\name{get_portfolios}
\alias{get_portfolios}
\title{Returns a specific Company object (Company Portfolio)}
\usage{
get_portfolios(tickers, portfolios)
}
\arguments{
\item{tickers}{A single ticker or vector of tickers that the user desires portfolios for.}

\item{portfolios}{A list of company portfolios, i.e. Company Portfolio objects.}
}
\description{
Given a list of tickers and a data frame of portfolios, returns
particular company portfolios based on inputted tickers.
}
\examples{
tickers <- c("AAPL","GOOG")
data(portfolios)
get_portfolios(tickers,portfolios)
}

