% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_qmj.R
\name{get_qmj}
\alias{get_qmj}
\title{Returns a qmj object for a specific company.}
\usage{
get_qmj(companies, ticker, financials, prices)
}
\arguments{
\item{companies}{A data frame of companies.}

\item{ticker}{A company ticker as a Character. Must already be present in the companies data frame.}

\item{financials}{A data frame containing financial information for the given companies.}

\item{prices}{A data frame containing the daily market closing prices and returns.}
}
\description{
Given a data frame of companies, a ticker, a data frame of financial statements, and a data frame of prices, creates
a particular company qmj object for the company of the corresponding ticker.
}
\examples{
data(companies)
ticker <- "AAPL"
data(financials)
data(prices)
get_qmj(companies,ticker,financials,prices)
}

