% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_portfolios.R
\name{get_qmjAnalyses}
\alias{get_qmjAnalyses}
\title{Returns specific qmjAnalysis objects in the existing data set of qmjAnalyses.}
\usage{
get_qmjAnalyses(tickers, qmjAnalyses)
}
\arguments{
\item{tickers}{A single ticker or vector of tickers that the user desires qmjAnalyses for.}

\item{qmjAnalyses}{A list of qmjAnalyses, i.e. qmjAnalysis objects.}
}
\description{
Given a vector of tickers and a data frame of qmjAnalyses, returns
particular company qmjAnalyses based on inputted tickers.
}
\examples{
tickers <- c("AAPL","GOOG")
data(qmjAnalyses)
get_qmjAnalyses(tickers,qmjAnalyses)
}

