% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_qmjs.R
\name{get_qmjs}
\alias{get_qmjs}
\title{Returns specific qmj objects from a list of qmj objects.}
\usage{
get_qmjs(tickers, qmjs)
}
\arguments{
\item{tickers}{A single ticker or vector of tickers that the user desires qmjs for.}

\item{qmjs}{A list of qmjs, i.e. qmj objects.}
}
\description{
Given a vector of tickers and a list of qmjs, returns
particular company qmjs based on inputted tickers.
}
\examples{
tickers <- c("AAPL","GOOG")
data(qmjs)
get_qmjs(tickers,qmjs)
}

