% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_raw_balancesheets.R
\name{get_raw_balancesheets}
\alias{get_raw_balancesheets}
\title{Gets raw balancesheets.}
\usage{
get_raw_balancesheets()
}
\description{
Retrieves raw data for balancesheets. Returns a large matrix, where
 individual elements correspond to a company's balancesheet statements
 over four years.
}

