% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_raw_cashflows.R
\name{get_raw_cashflows}
\alias{get_raw_cashflows}
\title{Gets raw cash flow statements.}
\usage{
get_raw_cashflows()
}
\description{
Retrieves raw data for cash flows. Returns a large matrix, where
 individual elements correspond to a company's cash flow statements
 over four years.
}

