% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_raw_financials.R
\name{get_raw_financials}
\alias{get_raw_financials}
\title{Gets raw financial statements.}
\usage{
get_raw_financials()
}
\description{
Retrieves raw data for financial statements as returned by the quantmod package.
Structured as a list of three elements, which each element contains either balance sheets,
cash flows, or income statements for all companies.
}

