% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_raw_incomestatements.R
\name{get_raw_incomestatements}
\alias{get_raw_incomestatements}
\title{Gets raw income statements.}
\usage{
get_raw_incomestatements()
}
\description{
Retrieves raw data for income statements. Returns a large matrix, where
 individual elements correspond to a company's income statements
 over four years.
}

