% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_raw_prices.R
\name{get_raw_prices}
\alias{get_raw_prices}
\title{Gets raw price data from the extdata folder of the package.}
\usage{
get_raw_prices()
}
\description{
Retrieves raw data for stock price returns and closing prices. Returns a data frame, where columns
are associated with a company's calculated price return and its recorded closing price for the date.
}

