% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getbalancesheets.R
\name{getbalancesheets}
\alias{getbalancesheets}
\title{getbalancesheets}
\usage{
getbalancesheets()
}
\description{
Retrieves data from companies.csv and writes associated balanceflow statements to 'balancesheets.csv'.
}

