% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getcashflows.R
\name{getcashflows}
\alias{getcashflows}
\title{getcashflows}
\usage{
getcashflows()
}
\description{
Retrieves data from companies.csv and writes associated cashflow statements to 'cashflows.csv'.
}

