% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getdailydata.R
\name{getdailydata}
\alias{getdailydata}
\title{getdailydata}
\usage{
getdailydata(x)
}
\arguments{
\item{x}{A dataframe of company names and tickers.}
}
\description{
Given a list of companies (names and tickers), writes .csv files for every company
storing price returns.
}

