% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getincomestatements.R
\name{getincomestatements}
\alias{getincomestatements}
\title{getincomestatements}
\usage{
getincomestatements()
}
\description{
Retrieves data from companies.csv and writes associated income statements to 'incomestatements.csv'.
}

