% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getinfo.R
\name{getinfo}
\alias{getinfo}
\title{getinfo}
\usage{
getinfo()
}
\description{
Retrieves data from the list of companies generated by
getcompanies() and writes associated cashflow statements,
income statements, and balance sheets into a data frame.
}
\examples{
\donttest{
getinfo()
}
}

