% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graph_portfolio.R
\name{graph_portfolio}
\alias{graph_portfolio}
\title{Graphical summary of quality scores}
\usage{
graph_portfolio(type = "marketcap")
}
\arguments{
\item{type}{Type that determines the graph created and the information presented.}
}
\description{
Returns a graphical summary of the portfolio. Be aware, missing values, or insufficient information, will result in some values
becoming 0.
}

