% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{growth}
\alias{growth}
\title{A dataframe of growth scores for companies listed in the Russell 3000}
\format{A data frame with 2999 rows and 7 variables:
\itemize{
   \item GPOA =  Four year growth in Gross Profits over Assets
   \item ROE = Four year growth in Return on Equity
   \item ROA = Four year growth in Return on Assets
   \item CFOA = Four year growth in Cash Flow over Assets
   \item GMAR = Four year growth in Gross Margin
   \item ACC = Four year growth in Accruals
 }}
\description{
Displays growth scores as well as its components for companies in the Russell 3000.
More information may be found on the technical vignette.
}
\keyword{data}

