% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{incomestatements}
\alias{incomestatements}
\title{Income statements of all companies}
\description{
A dataset containing all income statements, if available,
 for all companies in the companies.RData file. Among other things,
 stores:
 \itemize{
   \item Total Revenue.
   \item Gross Profit.
   \item Research & Development.
   \item Selling/General/Admin. Expenses, Total.
   \item Operating Income.
 }
}
\keyword{data}

