% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_yfinance.R
\name{install_yfinance}
\alias{install_yfinance}
\title{Install yfinance and dependencies}
\usage{
install_yfinance(..., envname = "r-qmj", new_env = identical(envname, "r-qmj"))
}
\arguments{
\item{...}{other arguments passed to [`reticulate::py_install()`]}

\item{envname}{The name, or full path, of the environment in which Python
packages are to be installed. When \code{NULL} (the default), the active
environment as set by the \code{RETICULATE_PYTHON_ENV} variable will be used;
if that is unset, then the \code{r-reticulate} environment will be used.}

\item{new_env}{Delete `envname` if it already exists}
}
\value{
No return value, called for installing Python yfinance module
}
\description{
`install_yfinance()` installs just the yfinance python package and it's
direct dependencies. Users may be asked to install miniconda for the python installations. 
Even if you first decline it, you can later install miniconda by running
[`reticulate::install_miniconda()`]
}
