% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/market_data.R
\name{market_data}
\alias{market_data}
\title{Produces component and quality scores when given relevant market data.}
\usage{
market_data(companies = qmjdata::companies,
  financials = qmjdata::financials, prices = qmjdata::prices)
}
\arguments{
\item{companies}{A data frame of company names and
tickers.}

\item{financials}{A data frame containing financial
information for the given companies.}

\item{prices}{A data frame containing the daily
market closing prices and returns.}
}
\value{
A data frame containing company names, tickers,
profitability z-scores, growth z-scores, safety z-scores,
payout z-scores, and quality z-scores. Organized by
quality in descending order.
}
\description{
Calculates market growth, payouts, safety, and
profitability of our list of companies for later
processing.
}
\details{
All parameters default to package data sets.
}
\examples{
## Rough structure I want this to be in:
## From beginnign to end.

## To immediately get quality scores using package data sets.

market_data()

## If we desire to produce a set of quality scores for a specific
## data frame of companies, which we'll call \\code{companies}

clean_downloads(companies)

raw_financials <- get_info(companies)
raw_prices <- get_prices(companies)

financials <- tidyinfo(raw_financials)
prices <- tidy_prices(raw_prices)

quality_scores <- market_data(companies, financials, prices)
}
\seealso{
\code{\link{market_profitability}}

\code{\link{market_growth}}

\code{\link{market_safety}}

\code{\link{market_payouts}}
}

