% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/market_payout.R
\name{market_payout}
\alias{market_payout}
\title{Collects payout z-scores for companies}
\usage{
market_payout(x, financials)
}
\arguments{
\item{x}{A dataframe of company names and tickers.}

\item{financials}{A dataframe containing financial statements for every company.}
}
\description{
Given a list of companies (names and tickers), a balance sheet, a cash flow statement,
and an income statement, calculates EISS, DISS, NPOP
and determines the z-score of overall payout based on the paper
Quality Minus Junk (Asness et al.) in Appendix page A3-4.
}
\examples{
data(companies)
data(financials)
x <- companies
market_payout(x, financials)
}

