% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/market_profitability.R
\name{market_profitability}
\alias{market_profitability}
\title{Collects profitability z-scores for companies}
\usage{
market_profitability(x, financials)
}
\arguments{
\item{financials}{A data frame containing financial statements for every company.}

\item{companies}{A data frame of company names and tickers.}
}
\description{
Given a data frame of companies (names and tickers) and a data frame of financial statements,
calculates GPOA, ROE, ROA, CFOA, GMAR, ACC and determines the z-score of overall profitability
for each company based on the paper Quality Minus Junk (Asness et al.) in Appendix page A2.
}
\examples{
data(companies)
data(financials)
market_profitability(companies, financials)
}

