% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{payouts}
\alias{payouts}
\title{A dataframe of payout scores for companies listed in the Russell 3000}
\description{
Displays payout scores as well as its components, this includes:
\itemize{
   \item EISS = Minus one-year percent change in split-adjusted number of shares.
   \item DISS = Minus one-year percent change in total debt.
   \item NPOP = Sum of total net payout over the past 4 years divided by total profits over the past 4 years.
 }
}
\details{
More information may be found on the Technical vignette.
Last updated: January 2015
}
\keyword{data}

