% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qmj-class.R
\docType{methods}
\name{plot_growth}
\alias{plot_growth}
\alias{plot_growth,qmj-method}
\title{Plots the relative growth of the qmj object.}
\usage{
plot_growth(theObject, growth_data_frame)
}
\value{
Displays a growth histogram, highlighting the bin containing the chosen qmj object/company.
}
\description{
Plots the relative growth of the qmj object.
}
\examples{
\dontrun{
data(qmjs)
data(growth)
plot_growth(qmjs[[1]], growth)
}
}

