% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qmj-class.R
\docType{methods}
\name{plot_payouts}
\alias{plot_payouts}
\alias{plot_payouts,qmj-method}
\title{Plots the relative payouts of the qmj object.}
\usage{
plot_payouts(theObject, payouts_data_frame)
}
\value{
Displays a payouts histogram, highlighting the bin containing the chosen qmj object/company.
}
\description{
Plots the relative payouts of the qmj object.
}
\examples{
\dontrun{
data(qmjs)
data(payouts)
plot_payouts(qmjs[[1]], payouts)
}
}

