% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qmj-class.R
\docType{methods}
\name{plot_profitability}
\alias{plot_profitability}
\alias{plot_profitability,qmj-method}
\title{Plots the relative profitability of the qmj object.}
\usage{
plot_profitability(theObject, profitability_data_frame)
}
\value{
Displays a profitability histogram, highlighting the bin containing the chosen qmj object/company.
}
\description{
Plots the relative profitability of the qmj object.
}
\examples{
\dontrun{
data(qmjs)
data(profitability)
plot_profitability(qmjs[[1]], profitability)
}
}

