% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qmj-class.R
\docType{methods}
\name{plot_safety}
\alias{plot_safety}
\alias{plot_safety,qmj-method}
\title{Plots the relative safety of the qmj object.}
\usage{
plot_safety(theObject, safety_data_frame)
}
\value{
Displays a safety histogram, highlighting the bin containing the chosen qmj object/company.
}
\description{
Plots the relative safety of the qmj object.
}
\examples{
\dontrun{
data(qmjs)
data(safety)
plot_safety(qmjs[[2]], safety)
}
}

