% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{portfolios}
\alias{portfolios}
\title{A list of Company Portfolios}
\description{
A list of company portfolios. These portfolios allow for much more in-depth analysis of a specific company
using the summarize, plot_quality, and view functions. (view_profitability, view_growth, view_safety, view_payouts)
Last updated: January 2015
}
\keyword{data}

