% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pricereturns.R
\name{pricereturns}
\alias{pricereturns}
\title{pricereturns}
\usage{
pricereturns(x)
}
\arguments{
\item{x}{An xts object with closing prices.}
}
\description{
Given an xts object with closing prices, returns an xts object with price returns.
}

