% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{prices}
\alias{prices}
\title{A dataframe of price returns and closing prices for companies}
\description{
Stores price returns and closing prices for the past two years (if available) for the Russell 3000 Index companies
as well as the S&P 500, to serve as a benchmark.
}
\keyword{data}

