% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{qmjs}
\alias{qmjs}
\title{A list of qmj objects.}
\description{
A list of qmj objects. These objects allow for much more in-depth analysis of a specific company
using the summarize, plot_quality, and view functions. (view_profitability, view_growth, view_safety, view_payouts)
Last updated: January 2015
}
\keyword{data}

