% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{safety}
\alias{safety}
\title{A dataframe of safety scores for companies listed in the Russell 3000}
\description{
Displays safety scores as well as its components, this includes:
\itemize{
   \item BAB = Beta, calculated by comparison with the S&P 500
   \item IVOL = Idiosyncratic Volatility, also calculated using the S&P 500
   \item LEV = Leverage
   \item O = Ohlson O-Score.
   \item Z = Altman Z-Score
   \item EVOL = Standard deviation of annual ROE over the past four years.
 }
}
\details{
More information may be found on the Technical vignette.
Last updated: January 2015
}
\keyword{data}

