% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/save_companies.R
\name{save_companies}
\alias{save_companies}
\title{Saves a data frame of companies to the package data set.}
\usage{
save_companies(companies)
}
\arguments{
\item{companies}{A dataframe a "tickers" column.}
}
\description{
Given a list of companies with "tickers" column, stores data as an RData file in the /data folder of qmj.
The list of companies you use MUST have a variable name of "companies".
}

