% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tidy_cashflows.R
\name{tidy_cashflows}
\alias{tidy_cashflows}
\title{tidy_cashflows}
\usage{
tidy_cashflows(x)
}
\description{
Tidies given cashflow data, formatted in typical style of quantmods viewFinancials function.
}

