% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_helper.R
\name{tidy_helper}
\alias{tidy_helper}
\title{Main helper function for all tidy functions.}
\usage{
tidy_helper(x)
}
\arguments{
\item{x}{A matrix containing financial information, 
either cash flows, balancesheets, or income statements,
downloaded from Google Finance. The formatting of the 
matrix has not been altered yet, as if just retrieved.}
}
\description{
This function does the main work of converting
raw financial data into organized data frames. 
It is used by qmj's tidy functions to reuse common
code and to avoid potential mistakes from repeating
similar processes.
}
\seealso{
\code{\link{tidy_cashflows}}

\code{\link{tidy_balancesheets}}

\code{\link{tidy_incomestatements}}
}
