% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/update_dailydata.R
\name{update_dailydata}
\alias{update_dailydata}
\title{Updates and tidies daily data.}
\usage{
update_dailydata(x)
}
\arguments{
\item{x}{A data frame of company names and tickers.}
}
\description{
Takes in a data frame of companies (must include "tickers" column), reads in the daily data and stores that information
in the /extdata folder, then tidies that raw data to be saved as "tidydaily" in the /data folder.
}

