% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qmj-class.R
\docType{methods}
\name{view_growth}
\alias{view_growth}
\alias{view_growth,qmj-method}
\title{Returns the growth of the qmj object.}
\usage{
view_growth(theObject)
}
\value{
A data frame with 1 row and 8 variables that describes growth of the qmj object.
}
\description{
Returns the growth of the qmj object.
}
\examples{
\dontrun{
qmjs <- data(qmjs)
view_growth(qmjs[[1]])
}
}

