% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qmj-class.R
\docType{methods}
\name{view_profitability}
\alias{view_profitability}
\alias{view_profitability,qmj-method}
\title{Returns the profitability of the qmj object.}
\usage{
view_profitability(theObject)

\S4method{view_profitability}{qmj}(theObject)
}
\value{
A data frame with 1 row and 8 variables that describes profitability of the qmj object.
}
\description{
Returns the profitability of the qmj object.
}
\examples{
\dontrun{
qmjs <- data(qmjs)
view_profitability(qmjs[[1]])
}
}

