% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qmj-class.R
\docType{methods}
\name{view_quality}
\alias{view_quality}
\alias{view_quality,qmj-method}
\title{Returns the quality of the qmj object.}
\usage{
view_quality(theObject)
}
\value{
A data frame with 1 row and 6 variables that describes the quality score of the qmj object.
}
\description{
Returns the quality of the qmj object.
}
\examples{
\dontrun{
qmjs <- data(qmjs)
view_quality(qmjs[[1]])
}
}

