% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qmj-class.R
\docType{methods}
\name{view_ticker}
\alias{view_ticker}
\alias{view_ticker,qmj-method}
\title{Return the ticker of the qmj object.}
\usage{
view_ticker(theObject)

\S4method{view_ticker}{qmj}(theObject)
}
\value{
The ticker of the qmj object as a character.
}
\description{
Return the ticker of the qmj object.
}
\examples{
\dontrun{
qmjs <- data(qmjs)
view_ticker(qmjs[[1]])
}
}

