% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/collect_market_safety.R
\name{collect_market_safety}
\alias{collect_market_safety}
\title{Collects safety z-scores for companies}
\usage{
collect_market_safety(x, BS, CF, IS, extrafin, daily)
}
\arguments{
\item{x}{A dataframe of company names and tickers.}

\item{BS}{A dataframe containing balance sheet information for every company.}

\item{CF}{A dataframe containing cash flow information for every company.}

\item{IS}{A dataframe containing income statement information for every company.}

\item{daily}{A dataframe containing the daily market closing prices and returns.}
}
\description{
Given a list of companies (names and tickers), a balance sheet, a cash flow statement,
and an income statement, calculates BAB, IVOL, LEV, O, Z, and EVOL.
and determines the z-score of overall profitability based on the paper
Quality Minus Junk (Asness et al.) in Appendix page A2.
}
\examples{
data(companies)
data(tidybalance)
data(tidycash)
data(tidyincome)
data(extrafin)
data(tidydaily)
x <- companies
BS <- tidybalance
CF <- tidycash
IS <- tidyincome
extrafin <- extrafin
daily <- tidydaily
collect_market_safety(x, BS, CF, IS, extrafin, daily)
}

