% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{companies}
\alias{companies}
\title{A list of all companies of interest}
\description{
Stores sample set of companies (namely, all those starting with the letter "A" as found on investorguide.com)
}
\keyword{data}

