% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_companies.R
\name{get_companies}
\alias{get_companies}
\title{Gets the company names and tickers from the most recent
Russell 3000 Index.}
\usage{
get_companies(x = system.file("extdata/companies.txt", package = "qmj"))
}
\arguments{
\item{x}{Specifies the filepath of the text file containing the company names and
tickers of interest.}
}
\description{
Reads in the contents of a text file created from the
pdf of company names and tickers given by the Russell
3000 Index. Because there is not a consistent way to read
pdfs into text files using R across different operating
systems, the user must copy and paste the contents of the
Russell 3000 Index into a text file for this function to
process the data correctly. This can be done simply by
selecting all on the Russell 3000 Index and pasting into
an empty document with the .txt extension. The contents
of the Russell 3000 Index can be found
\href{https://www.russell.com/documents/indexes/membership/membership-russell-3000.pdf}{here}.
}
\details{
If you wish to create your own text files of companies for get_companies to process,
please be sure that each letter of all company names and tickers is capitalized and
each word is separated by a single space, with each line starting with the company name
followed by the company ticker. For example, if you want to put two companies in your text
file with the first having the name Apple and ticker AAPL and the second having the name
Google and ticker GOOG, please format the text file as follows:

APPLE AAPL \cr
GOOGLE GOOG

Once you have your text file, you must pass the filepath as an argument to get_companies.
For example, if your text file is called "companies.txt", you would use
\code{get_companies("companies.txt")}. get_companies by default uses a text file
in the package, so \code{get_companies()} will produce the data frame of names and tickers
for the most recent Russell 3000 Index already.
}
\examples{
get_companies()
}

