% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_companies.R
\name{get_companies}
\alias{get_companies}
\title{Gets the company names and tickers from the most recent Russell 3000 Index.}
\usage{
get_companies()
}
\description{
Reads in the contents of a text file created from the pdf of company
names and tickers given by the Russell 3000 Index.
}
\examples{
get_companies()
}

