% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info}
\alias{get_info}
\title{Gets raw company balance sheets, income statements, and
cash flows from Google Finance.}
\usage{
get_info(x = qmjdata::companies)
}
\arguments{
\item{x}{A data frame of companies. Must have a ticker column.}
}
\description{
\code{get_info} grabs financial data for a given data frame of companies.
}
\details{
For each ticker in the data frame of companies, \code{get_info} grabs
financial data and generates a list with three elements.

Each element is a large list
containing all the balance sheets, income statements, or cash
flow statements for all companies. Also writes .RData files
for every company in the /extdata folder in the package folder.
If canceled partway through, function is able to find and
re-read this data, allowing resumption of progress.
}
\examples{
sub_comps <- qmjdata::companies[1:2,]
get_info(sub_comps)
}
\seealso{
\code{\link{get_prices}}
}

