% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info}
\alias{get_info}
\title{Gets company balance sheets, income statements, and cash flows.}
\usage{
get_info(x)
}
\arguments{
\item{companies}{A data frame of companies. Must have a ticker column.}
}
\description{
Retrieves data for the data frame of companies generated by
getcompanies() and writes associated cashflow statements,
income statements, and balance sheets into a data frame.
}
\examples{
\dontrun{
data(companies)
get_info(companies)
}
}

