% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_portfolio.R
\name{get_portfolio}
\alias{get_portfolio}
\title{Summary of quality scores}
\usage{
get_portfolio(top = 5, bottom = 5)
}
\arguments{
\item{top}{The number of top ranked companies to be returned.}

\item{bottom}{The number of bottom ranked companies to be returned.}
}
\description{
Returns, by default, the top five and the bottom five companies ordered by
calculated quality score. Also returns the standard deviation of quality scores
and the interquartile range.
}

