% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_portfolios.R
\name{get_portfolios}
\alias{get_portfolios}
\title{Returns a specific Company object (Company Portfolio)}
\usage{
get_portfolios(tickers, portfolios)
}
\arguments{
\item{tickers}{A single ticker or vector of tickers that the user desires portfolios for.}

\item{portfolios}{A list of company portfolios, i.e., Company objects.}
}
\description{
Given a data frame of companies, a ticker, a data frame with financial statements, and a data frame of prices, creates
a particular company portfolio and returns the result.
}

