% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getincomestatements.R
\name{getincomestatements}
\alias{getincomestatements}
\title{getincomestatements}
\usage{
getincomestatements()
}
\description{
Retrieves data from the list of companies generated by getcompanies() and
writes associated income statements into a data frame.
}
\examples{
getincomestatements()
}

