% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{growth}
\alias{growth}
\title{A dataframe of growth scores for companies listed in the Russell 3000}
\description{
Displays growth scores as well as its components, this includes:
\itemize{
   \item GPOA =  Four year growth in Gross Profits over Assets
   \item ROE = Four year growth in Return on Equity
   \item ROA = Four year growth in Return on Assets
   \item CFOA = Four year growth in Cash Flow over Assets
   \item GMAR = Four year growth in Gross Margin
   \item ACC = Four year growth in Accruals
 }
}
\details{
More information may be found on the Technical vignette.
Last updated: January 2015
}
\keyword{data}

