% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/market_data.R
\name{market_data}
\alias{market_data}
\title{Collects all relevant market data (growth, payout, profitability, and safety) and orders companies
by descending quality.}
\usage{
market_data(x, financials, daily)
}
\arguments{
\item{x}{A dataframe of company names and tickers.}

\item{financials}{a formatted data frame containing financial information for the given companies.}

\item{daily}{A dataframe containing the daily market closing prices and returns.}
}
\description{
Calculates market growth, payouts, safety, and profitability of our list of companies
for later processing.
}
\examples{
data(companies)
data(financials)
market_data(companies, financials)
}

