% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/market_growth.R
\name{market_growth}
\alias{market_growth}
\title{Collects growth z-scores for companies}
\usage{
market_growth(companies = qmjdata::companies,
  financials = qmjdata::financials)
}
\arguments{
\item{companies}{A data frame of company names and tickers.}

\item{financials}{A data frame containing financial
statements for every company.}
}
\description{
Given a data frame of companies (names and tickers) and
a data frame of financial statements, calculates GPOA,
ROE, ROA, CFOA, GMAR, ACC over a four-year time span
and determines the z-score of overall growth for each
company based on the paper Quality Minus Junk
(Asness et al.) in Appendix page A2.
}
\examples{
companies <- qmjdata::companies[50:51,]
market_growth(companies, qmjdata::financials)
}
\seealso{
\code{\link{market_data}}

\code{\link{market_profitability}}

\code{\link{market_safety}}

\code{\link{market_payouts}}
}

