% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/market_payout.R
\name{market_payout}
\alias{market_payout}
\title{Collects payout z-scores for companies}
\usage{
market_payout(x, financials)
}
\arguments{
\item{financials}{A data frame containing financial statements for every company.}

\item{companies}{A data frame of company names and tickers.}
}
\description{
Given a data frame of companies (names and tickers) and a data frame of financial statements,
calculates EISS, DISS, NPOP and determines the z-score of overall payout for each company based on the paper
Quality Minus Junk (Asness et al.) in Appendix page A3-4.
}
\examples{
data(companies)
data(financials)
market_payout(companies, financials)
}

