% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/market_payouts.R
\name{market_payouts}
\alias{market_payouts}
\title{Collects payout z-scores for companies}
\usage{
market_payouts(companies = qmjdata::companies,
  financials = qmjdata::financials)
}
\arguments{
\item{companies}{A data frame of company names and tickers.}

\item{financials}{A data frame containing financial statements for every
company.}
}
\description{
Given a data frame of companies (names and tickers) and a data frame of
financial statements, calculates EISS, DISS, NPOP and determines the
z-score of overall payout for each company based on the paper Quality
Minus Junk (Asness et al.) in Appendix page A3-4.
}
\examples{
companies <- qmjdata::companies[50:51,]
market_payouts(companies, qmjdata::financials)
}
\seealso{
\code{\link{market_data}}

\code{\link{market_profitability}}

\code{\link{market_growth}}

\code{\link{market_safety}}
}

