% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/market_profitability.R
\name{market_profitability}
\alias{market_profitability}
\title{Collects profitability z-scores for companies}
\usage{
market_profitability(x, financials)
}
\arguments{
\item{x}{A dataframe of company names and tickers.}

\item{financials}{A dataframe containing financial statements for every company.}
}
\description{
Given a list of companies (names and tickers), a balance sheet, a cash flow statement,
and an income statement, calculates GPOA, ROE, ROA, CFOA, GMAR, ACC,
and determines the z-score of overall profitability based on the paper
Quality Minus Junk (Asness et al.) in Appendix page A2.
}
\examples{
data(companies)
data(financials)
x <- companies
market_profitability(x, financials)
}

