% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/market_safety.R
\name{market_safety}
\alias{market_safety}
\title{Collects safety z-scores for companies}
\usage{
market_safety(x, financials, daily)
}
\arguments{
\item{x}{A dataframe of company names and tickers.}

\item{financials}{A dataframe containing financial statements for every company.}

\item{daily}{A dataframe containing the daily market closing prices and returns.}
}
\description{
Given a list of companies (names and tickers), a balance sheet, a cash flow statement,
and an income statement, calculates BAB, IVOL, LEV, O, Z, and EVOL.
and determines the z-score of overall profitability based on the paper
Quality Minus Junk (Asness et al.) in Appendix page A2.
}
\examples{
data(companies)
data(financials)
data(extrafin)
data(tidydaily)
x <- companies
financials <- financials
daily <- tidydaily
market_safety(x, financials, daily)
}

