% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/market_safety.R
\name{market_safety}
\alias{market_safety}
\title{Collects safety z-scores for companies}
\usage{
market_safety(x, financials, daily)
}
\arguments{
\item{financials}{A data frame containing financial statements for every company.}

\item{companies}{A data frame of company names and tickers.}

\item{prices}{A data frame containing the daily market closing prices and returns.}
}
\description{
Given a data frame of companies (names and tickers), a data frame of financial
statements, and a data frame of daily price data, calculates BAB, IVOL, LEV, O, Z, and EVOL
and determines the z-score of overall safety for each company based on the paper
Quality Minus Junk (Asness et al.) in Appendix page A2.
}
\examples{
data(companies)
data(financials)
data(prices)
x <- companies[50:51,]
daily <- prices
market_safety(x, financials, daily)
}
\seealso{
\code{\link{market_data}}

\code{\link{market_profitability}}

\code{\link{market_growth}}

\code{\link{market_payouts}}
}

