% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prices.R
\docType{data}
\name{prices}
\alias{prices}
\title{A dataframe of price returns and closing prices for companies}
\format{A data frame with 1,475,934 rows and 4 variables}
\source{
Google Finance, accessed through quantmod
}
\description{
Stores price returns and closing prices for the past two years (if available) for the Russell 3000 Index companies
as well as the S&P 500 (uniquely taken from Yahoo finance), to serve as a benchmark.
}
\keyword{data}

