% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{profitability}
\alias{profitability}
\title{A dataframe of profitability scores for companies listed in the Russell 3000}
\format{A data frame with 2999 rows and 7 variables:
\itemize{
   \item GPOA = Gross Profits over Assets
   \item ROE = Return on Equity
   \item ROA = Return on Assets
   \item CFOA = Cash Flow over Assets
   \item GMAR = Gross Margin
   \item ACC = Accruals
 }}
\description{
Displays profitability scores as well as its components for companies in the Russell 3000.
More information may be found on the technical vignette.
}
\details{
Last updated: January 2015
}
\keyword{data}

