% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qmj-package.R
\docType{package}
\name{qmj-package}
\alias{qmj-package}
\title{Exploring a quality minus junk approach to evaluating stocks}
\description{
The qmj package calculates quality scores for the companies in the Russell 3000 Index based
on the paper \emph{Quality Minus Junk} by Clifford Asness, Andrea Frazzini, and Lasse Pedersen.
}
\details{
Quality is a measure of a company's profitability, growth, safety, and payouts. By using publicly
available data for company balance sheets, income statements, and cash flows, the qmj package
calculates relative quality z-scores for companies. Utility functions in the class qmj also allow
users to further analyze the quality of stocks.

All functions and datasets are documented, and are freely available for use.
Index of datasets:
\itemize{
\item companies - all publicly traded companies in the Russell 3000 Index
\item financials - financial statements for companies in the companies dataset
\item profitability - most recently measured profitability z-scores and their components
\item growth - most recently measured growth z-scores and their components
\item safety - most recently measured safety z-scores and their components
\item payouts - most recently measured payouts z-scores and their components
\item quality - most recently measure quality z-scores and their components
\item qmjs - objects of class qmj, providing utility functions for further analysis of company quality
\item prices - daily prices and price returns for the past two years
}
}
\references{
Asness, Clifford S., Andrea Frazzini, and Lasse H. Pedersen. "Quality Minus Junk." AQR (2013)
}

