% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qmj.R
\docType{package}
\name{qmj}
\alias{qmj}
\alias{qmj-package}
\title{Quality Minus Junk}
\description{
Calculates and manipulates data on the quality of companies.
}
\details{
This package is currently still in early development.
}

