% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qmj-class.R, R/qmj.R
\docType{class}
\name{qmj}
\alias{qmj}
\alias{qmj-class}
\alias{qmj-package}
\title{qmj}
\description{
The qmj object contains all details related to a company's quality score,
and also provides some tools for analyzing a specific company in more detail.

Calculates and manipulates data on the quality of companies.
}
\details{
\section{Slots}{
   \describe{
     \item{\code{ticker}:}{Object of class \code{"character"}.}
     \item{\code{pGPOA}:}{Object of class \code{"numeric"}.}
     \item{\code{pROE}:}{Object of class \code{"numeric"}.}
     \item{\code{pROA}:}{Object of class \code{"numeric"}.}
     \item{\code{pCFOA}:}{Object of class \code{"numeric"}.}
     \item{\code{pACC}:}{Object of class \code{"numeric"}.}
     \item{\code{growth}:}{Object of class \code{"numeric"}.}
     \item{\code{gGPOA}:}{Object of class \code{"numeric"}.}
     \item{\code{gROE}:}{Object of class \code{"numeric"}.}
     \item{\code{gROA}:}{Object of class \code{"numeric"}.}
     \item{\code{gCFOA}:}{Object of class \code{"numeric"}.}
     \item{\code{gGMAR}:}{Object of class \code{"numeric"}.}
     \item{\code{gACC}:}{Object of class \code{"numeric"}.}
     \item{\code{safety}:}{Object of class \code{"numeric"}.}
     \item{\code{sBAB}:}{Object of class \code{"numeric"}.}
     \item{\code{sIVOL}:}{Object of class \code{"numeric"}.}
     \item{\code{sLEV}:}{Object of class \code{"numeric"}.}
     \item{\code{sO}:}{Object of class \code{"numeric"}.}
     \item{\code{sZ}:}{Object of class \code{"numeric"}.}
     \item{\code{sEVOL}:}{Object of class \code{"numeric"}.}
     \item{\code{payouts}:}{Object of class \code{"numeric"}.}
     \item{\code{pEISS}:}{Object of class \code{"numeric"}.}
     \item{\code{pDISS}:}{Object of class \code{"numeric"}.}
     \item{\code{pNPOP}:}{Object of class \code{"numeric"}.}
     \item{\code{quality}:}{Object of class \code{"numeric"}.}
   }
 }

*This package is currently still in early development.* \cr
  RECOMMENDATION FOR EARLY USE: \cr
  \itemize{
  \item All functions and datasets are documented, and are freely available for use.
  \item getcompanies() is an alternative way of accessing the companies.RData data.
  \item collectmarketdata() provides an early look into *very rough* initial calculations
  for quality in the sample companies.
  \item *Warning*: Do not call getcashflows, getincomestatements, or getbalancesheets.
  While these functions correctly work, the functions do require several minutes to
  fully finish downloading the relevant financial documents from the web, returning
  a large matrix containing the desired documents.
  }
}

