% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{quality}
\alias{quality}
\title{A dataframe quality scores for companies listed in the Russell 3000}
\description{
Displays overall quality scores as well as the scores for profitability, growth,
safety, and payouts.
}
\details{
Last updated: January 2015
}
\keyword{data}

