% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{quality}
\alias{quality}
\title{A dataframe of quality scores for companies listed in the Russell 3000}
\description{
Displays overall quality scores as well as the scores for profitability, growth,
safety, and payouts. Companies are sorted in order of quality score, with NAs stored
at the end of the data set.
}
\details{
If partial information exists (i.e., a profitability score
was able to be calculated), then those scores are kept for that company, even if
insufficient information exists to produce a quality score.

Last updated: January 2015
}
\keyword{data}

