% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tidy_balancesheets.R
\name{tidy_balancesheets}
\alias{tidy_balancesheets}
\title{Makes raw balancesheet data usable and readable.}
\usage{
tidy_balancesheets(x)
}
\arguments{
\item{x}{A list of raw cash flow data produced from quantmod}
}
\value{
Returns a data set that's been "tidied" up for use
by other functions in this package.
}
\description{
Processes raw balance sheet data produced from quantmod
into a tidy data frame. Raw balance sheet data must be
formatted in a list such that every element is a data
frame or matrix containing quantmod data.
}
\details{
\code{tidy_balancesheets} produces a data frame that is
"tidy" or more readily readable by a user and usable by
other functions within this package.
}
\examples{
sub_comps <- qmjdata::companies[1:2,]
raw_data <- get_info(sub_comps)
tidybalance <- tidy_balancesheets(raw_data[[3]])
}
\seealso{
\code{\link{get_info}}

\code{\link{tidyinfo}}

\code{\link{tidy_cashflows}}

\code{\link{tidy_incomestatements}}
}

