% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tidy_balancesheets.R
\name{tidy_balancesheets}
\alias{tidy_balancesheets}
\title{tidy_balancesheets}
\usage{
tidy_balancesheets(x)
}
\description{
Tidies given balancesheet data, formatted in typical style of quantmods viewFinancials function.
}

