% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tidy_cashflows.R
\name{tidy_cashflows}
\alias{tidy_cashflows}
\title{Makes raw cash flow data usable and readable.}
\usage{
tidy_cashflows(x)
}
\arguments{
\item{x}{A list of raw cash flow data produced from quantmod}
}
\value{
Returns a data set that's been 'tidied' up for use
by other functions in this package.
}
\description{
Processes raw cash flow data from quantmod to return
a tidied data frame. Raw cash flow data must be formatted
in a list such that every element is a data frame or
matrix containing quantmod data.
}
\details{
\code{tidy_cashflows} produces a data frame that is 'tidy'
or more readily readable by a user and usable by other
functions within this package.
}
\examples{
sub_comps <- qmjdata::companies[1:2,]
raw_data <- get_info(sub_comps)
tidycash <- tidy_cashflows(raw_data[[1]])
}
\seealso{
\code{\link{get_info}}

\code{\link{tidyinfo}}

\code{\link{tidy_balancesheets}}

\code{\link{tidy_incomestatements}}
}

