% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tidy_dailydata.R
\name{tidy_dailydata}
\alias{tidy_dailydata}
\title{tidy_dailydata}
\usage{
tidy_dailydata(x)
}
\arguments{
\item{x}{The resultant data frame from calling getdailydata(x).}
}
\description{
Given the data frame organized similarly to that produced by getdailydata()
 in the qmj library, tidies the data and returns the resultant data frame.
}

