% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tidy_incomestatements.R
\name{tidy_incomestatements}
\alias{tidy_incomestatements}
\title{tidy_incomestatements}
\usage{
tidy_incomestatements(x)
}
\description{
Tidies given income statement data, formatted in typical style of quantmods viewFinancials function.
}

