% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tidy_incomestatements.R
\name{tidy_incomestatements}
\alias{tidy_incomestatements}
\title{Makes raw incomestatement data usable and readable.}
\usage{
tidy_incomestatements(x)
}
\arguments{
\item{x}{A list of raw incomestatement file data produced from quantmod}
}
\value{
Returns a data set that's been "tidied" up for use by other
functions in this package.
}
\description{
Tidies raw income statement data produced from quantmod and returns
the tidied data frame. Raw income statement data must be formatted
in a list such that every element is a data frame or matrix containing
quantmod data.
}
\details{
\code{tidy_incomestatements} produces a data frame that is "tidy"
or more readily readable by a user and usable by other functions within
this package.
}
\examples{
sub_comps <- qmjdata::companies[1:2,]
raw_data <- get_info(sub_comps)
tidyincome <- tidy_incomestatements(raw_data[[2]])
}
\seealso{
\code{\link{get_info}}

\code{\link{tidyinfo}}

\code{\link{tidy_cashflows}}

\code{\link{tidy_balancesheets}}
}

