% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tidy_prices.R
\name{tidy_prices}
\alias{tidy_prices}
\title{Tidies raw prices data set}
\usage{
tidy_prices(x)
}
\arguments{
\item{x}{Raw daily data, as produced by get_prices()}
}
\description{
Tidies raw prices and returns a tidied, usable data frame. Raw data should be structured identically to that produced
by get_prices(), as this function depends on that structure.
}

