% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tidyinfo.R
\name{tidyinfo}
\alias{tidyinfo}
\title{Tidies raw financial data.}
\usage{
tidyinfo(x)
}
\arguments{
\item{x}{A list of lists of financial statements. Generated from get_info(companies).}
}
\description{
Reads in raw financial data in the /extdata folder of the qmj package,
tidies the data found, and then saves that data in the /data folder.
}

